v1.0
First version of a working A* pathfinding system that works on an agent to track a Transform down and move to it with various settings. I also can now use layers to make certein terrain more difficult to move trough for the agent, so it chooses a different path. Optimizing coming soon